+++ CHANGELOG.md
@@ -12,6 +12,9 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Added
 
+- New commands to view details for user risk profiles:
+  - `code42 users list-risk-profiles`
+  - `code42 users show-risk-profile`
 - Proxy support via `HTTPS_PROXY` environment variable.
 
 ## 1.15.0 - 2022-08-23